requirements.txt
tqdm
requests
kagglehub
numpy
pandas>=2.0.0
matplotlib
seaborn
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai
langchain-community>=0.0.10
langchain-chroma
transformers
gradio
notebook>=7.0.0
ipywidgets
sentence-transformers>=2.2.2
chromadb>=0.4.22